requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
jdcal==1.4
Markdown==3.0.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
six==1.12.0
tablib==0.13.0
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
